Major bugfix in finestructure estimates.
This seems to make whalen and mixsea versions agree.

In the spectral integration I was using psd[i0:idx] which is [i0, idx)
half-closed interval. But in the next line idxint was [i0, idx] closed
interval, so there was a mismatch which was screwing things up.

When used on NATRE microstrucutre, there's a difference between Whalen &
Kunze that is plausibly the Rω value being 3 or 7.